test: remove `test_bugs.py` and move all its tests
