+++ test/test_lopf_global_constraints.py
@@ -2,7 +2,10 @@
 #
 # SPDX-License-Identifier: MIT
 
+import pandas as pd
 import pytest
+
+import pypsa
 
 
 def test_operational_limit_n_ac_dc_meshed(ac_dc_network):
@@ -138,3 +141,53 @@ def test_custom_variable_warnings(ac_dc_network, caplog):
             "could not be mapped to the network component because the component 'NonExistentComponent' does not exist",
         ]
     )
+
+
+def test_1449():
+    """
+    Inactive components with global carrier constraint should not break.
+    See https://github.com/PyPSA/PyPSA/issues/1449.
+    """
+    snapshots = pd.date_range("2025-01-01", freq="h", periods=5)
+
+    n = pypsa.Network(snapshots=snapshots)
+
+    n.add("Carrier", name="gird_with_emissions", co2_emissions=0.5)
+
+    n.add("Bus", name="b_electricity", unit="MW")
+
+    n.add(
+        "Generator",
+        name="grid",
+        bus="b_electricity",
+        p_nom=10,
+        marginal_cost=5,
+    )
+
+    n.add(
+        "Generator",
+        name="grid_2",
+        bus="b_electricity",
+        carrier="gird_with_emissions",
+        p_nom_extendable=True,
+        marginal_cost=5,
+        active=False,
+    )
+
+    n.add(
+        "Load",
+        name="demand",
+        bus="b_electricity",
+        p_set=pd.Series([0, 2, 0, 0, 0], index=snapshots),
+    )
+
+    n.add(
+        "GlobalConstraint",
+        name="CO2_limit",
+        carrier_attribute="co2_emissions",
+        sense="<=",
+        constant=50,
+    )
+
+    status, _ = n.optimize()
+    assert status == "ok"
